requirements.txt
apache-airflow==2.7.1
apache-airflow-providers-postgres
psycopg2-binary
requests
python-dotenv